+++ apps/src/CDOIDE/FileBrowser/styles/fileBrowser.css
