+++ apps/src/CDOIDE/Search/styles/search.css
