+++ apps/src/CDOIDE/SideBar/styles/sideBar.css
