+++ apps/src/CDOIDE/constants.ts
