+++ apps/src/CDOIDE/utils/getEmptyEditor.ts
