+++ apps/src/CDOIDE/utils/getEmptyProject.ts
