+++ apps/webpack.config.js
@@ -213,7 +213,7 @@ const WEBPACK_BASE_CONFIG = {
       repl: p('src/noop'),
       '@cdo/storybook': p('.storybook'),
       serialport: false,
-      '@cdoide': p('src/weblab2/CDOIDE'),
+      '@cdoide': p('src/CDOIDE'),
       '@cdo/generated-scripts': p('generated-scripts'),
     },
   },
